Fixed problem with computation of inverse mass matrix in Navier-Stokes preconditioner that -- somewhat bizarrely only showed in pseudo-elastic collapsible tube code when run on four or more processors!
Also checked in Richard's final version of the
three parallel block preconditioner tutorials.

Fixed a problem in setting TrilinosMLPrecondioner
(smoothing factor is a double rather than an int).



git-svn-id: svn+ssh://vummath.ma.man.ac.uk/home/svn-oomphlib/public/trunk@162 595e1d82-adf1-4282-bac7-b6073fa2ca8e

+++ src/generic/trilinos_preconditioners.h
@@ -203,7 +203,7 @@ namespace oomph
     }
 
    /// Function to set Smoother_damping
-   void set_smoother_damping(int smoother_damping) 
+   void set_smoother_damping(double smoother_damping) 
     {
      ML_parameters.set("smoother: damping factor", smoother_damping);
     }
